fix code

+++ test/arithmetic_test.cpp
@@ -10,7 +10,7 @@
 using namespace cyy::math;
 
 TEST_CASE("exponent") {
-  CHECK_EQ(exponent(integer(2), (uint32_t)64) - 1, integer(UINT64_MAX));
+  CHECK_EQ(exponent(integer(2), 64) - 1, integer(UINT64_MAX));
 }
 TEST_CASE("factorial") {
   integer a = 2;
